+++ NEWS.md
@@ -1,4 +1,6 @@
-# statsExpressions 0.2.0.9000
+# statsExpressions 0.2.1.9000
+
+# statsExpressions 0.2.1
 
 NEW FEATURES